% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_continuous.R
\name{check_continuous}
\alias{check_continuous}
\title{Check if the dates in a data frame are continuous.}
\usage{
check_continuous(data, dates)
}
\arguments{
\item{data}{A data.frame, or data frame extension (e.g. a tibble).}

\item{dates}{Column representing the date of the temperature measurements.}
}
\value{
A message indicating if the dates are continuous or if there are breaks. If
there are breaks, a vector of row numbers where the breaks occur is returned.
}
\description{
Check if the dates in a data frame are continuous.
}
\examples{
library(hatchR)
check_continuous(crooked_river, date)
}
