% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crooked_river.R
\docType{data}
\name{crooked_river}
\alias{crooked_river}
\title{Example dataset: Crooked River, Idaho}
\format{
Data frame with columns
\describe{
\item{date}{Date and time measurement was taken.}
\item{temperture}{Temperature in degrees Celsius.}
}
}
\source{
Dan Isaak, US Forest Service Rocky Mountain Research Station.
}
\usage{
crooked_river
}
\description{
Daily stream temperature data from Crooked River, a tributary to the
North Fork Boise River, Idaho.
}
\examples{
crooked_river
}
\keyword{datasets}
