% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatchR-package.R
\docType{package}
\name{hatchR-package}
\alias{hatchR}
\alias{hatchR-package}
\title{hatchR: Predict Fish Hatch and Emergence Timing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Predict hatch and emergence timing for a wide range of wild fishes using the effective value framework (Sparks et al., (2019) \doi{10.1139/cjfas-2017-0468}). 'hatchR' offers users access to established phenological models and the flexibility to incorporate custom parameterizations using external datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bmait101.github.io/hatchR/}
  \item \url{https://github.com/bmait101/hatchR}
  \item Report bugs at \url{https://github.com/bmait101/hatchR/issues}
}

}
\author{
\strong{Maintainer}: Bryan M. Maitland \email{bryan.maitland@usda.gov} (\href{https://orcid.org/0000-0002-4491-5064}{ORCID})

Authors:
\itemize{
  \item Morgan M. Sparks \email{morgan.sparks@usda.gov} (\href{https://orcid.org/0000-0003-0787-2218}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Eli Felts \email{eli_felts@fws.gov} (\href{https://orcid.org/0000-0002-5888-9266}{ORCID}) [contributor]
}

}
\keyword{internal}
