% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idaho.R
\docType{data}
\name{idaho}
\alias{idaho}
\title{Central Idaho Water Temperature Data}
\format{
Data frame
\describe{
\item{date}{Date measurement was taken.}
\item{site}{Site ID.}
\item{temp_c}{Temperature in degrees Celsius.}
}
}
\source{
https://hess.copernicus.org/articles/22/6225/2018/
}
\usage{
idaho
}
\description{
Water temperature data from Isaak et al. (2018).  Generally covers the
Boise, Payette, Clearwater, and upper Salmon River watersheds.
}
\examples{
idaho
}
\references{
Isaak, D.J., Luce, C.H., Chandler, G.L., Horan, D.L., Wollrab, S.P. (1990).
Principal components of thermal regimes in mountain river networks.
\emph{Hydrology and Earth System Sciences},
\bold{22(12)}, 6225--6240.
}
\keyword{datasets}
