% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table.R
\docType{data}
\name{model_table}
\alias{model_table}
\title{Table of phenology models}
\format{
Data frame
\describe{
\item{author}{Short article citation}
\item{species}{Species name}
\item{model}{Model number from B&M}
\item{dev.type}{Hatch or emergence}
\item{func}{character string desciption of the function}
}
}
\source{
To come...
}
\usage{
model_table
}
\description{
Description to come...
}
\examples{
model_table
}
\keyword{datasets}
