% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMRS_MT103.R
\docType{data}
\name{RMRS_MT103}
\alias{RMRS_MT103}
\title{Daily stream temperatures data}
\format{
Data frame with columns
\describe{
\item{SampleDate}{Date measurement was taken.}
\item{site}{Site ID.}
\item{temperture}{Temperature in degrees Celsius.}
}
}
\source{
Dan Isaak, US Forest Service Rocky Mountain Research Station
}
\usage{
RMRS_MT103
}
\description{
Daily stream temperature data from the US Forest Service Rocky
Mountain Research Station, Water and Watersheds Program.
}
\examples{
  RMRS_MT103
}
\keyword{datasets}
