% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_temp.R
\name{summarize_temp}
\alias{summarize_temp}
\title{Summarize temperature data to daily values}
\usage{
summarize_temp(data, dates, temperature)
}
\arguments{
\item{data}{A data.frame, or data frame extension (e.g. a tibble).}

\item{dates}{Column representing the date of temperature measurements.}

\item{temperature}{Column representing temperature values.}
}
\value{
A data.frame with summarized daily temperature values.
}
\description{
The \code{summarize_temp} function is used to summarize sub-daily temperature
measurements to obtain mean daily temperature.
}
\examples{
library(hatchR)
summarize_temp(
  data = idaho,
  dates = date,
  temperature = temp_c
)
}
