% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woody_island.R
\docType{data}
\name{woody_island}
\alias{woody_island}
\title{Example dataset: Woody Island, Lake Iliamna, Alaska}
\format{
Data frame with columns:
\describe{
\item{date}{Date measurement was taken.}
\item{temp_c}{Temperature in degrees Celsius.}
}
}
\source{
Sparks et al. (2019)
}
\usage{
woody_island
}
\description{
Temperature Data from Woody Island in Lake Iliamna, Alaska
}
\examples{
woody_island
}
\references{
Sparks, M.M., Falke, J.A., Quinn, T.A., Adkinson, M.D.,
Schindler, D.E., Bartz, K., Young, D., Westley, P.A.H. (2019). Influences of spawning timing, water temperature,
and climatic warming on early life history phenology in western
Alaska sockeye salmon.
\emph{Canadian Journal of Fisheries and Aquatic Sciences},
\bold{76(1)}, 123--135.
}
\keyword{datasets}
