% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crooked_river.R
\docType{data}
\name{crooked_river}
\alias{crooked_river}
\title{Daily stream temperatures data}
\format{
Data frame with columns
\describe{
\item{SampleDate}{Date measurement was taken.}
\item{site}{Site ID.}
\item{temperture}{Temperature in degrees Celsius.}
}
}
\source{
Dan Isaak, US Forest Service Rocky Mountain Research Station
}
\usage{
crooked_river
}
\description{
Daily stream temperature data from the US Forest Service Rocky
Mountain Research Station, Water and Watersheds Program.
}
\examples{
crooked_river
}
\keyword{datasets}
