% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatchR-package.R
\docType{package}
\name{hatchR-package}
\alias{hatchR}
\alias{hatchR-package}
\title{hatchR: Predict Fish Hatching and Emergence from Daily Temperature Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

`hatchR` is intended to bridge the analytic gap of taking statistical models developed in hatchery settings (e.g., Beacham and Murray 1990) and applying them to real world temperature data sets using an effective value framework developed by Sparks et al. (2019).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bmait101.github.io/hatchR/}
  \item \url{https://github.com/bmait101/hatchR}
  \item Report bugs at \url{https://github.com/bmait101/hatchR/issues}
}

}
\author{
\strong{Maintainer}: Bryan M. Maitland \email{bryan.maitland@usda.gov} (\href{https://orcid.org/0000-0002-4491-5064}{ORCID})

Authors:
\itemize{
  \item Morgan M. Sparks \email{morgan.sparks@usda.gov} (\href{https://orcid.org/0000-0003-0787-2218}{ORCID})
}

Other contributors:
\itemize{
  \item Eli Felts \email{eli_felts@fws.gov} (\href{https://orcid.org/0000-0002-5888-9266}{ORCID}) [contributor]
}

}
\keyword{internal}
