% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table.R
\docType{data}
\name{model_table}
\alias{model_table}
\title{Table of phenology models}
\format{
Data frame
\describe{
\item{author}{Short article citation}
\item{species}{Species name}
\item{model}{Model number from B&M}
\item{dev.type}{Hatch or emergence}
\item{func}{character string desciption of the function}
}
}
\source{
To come...
}
\usage{
model_table
}
\description{
Description to come...
}
\examples{
model_table
}
\references{
Beacham, T.D., Murray, C.B. (1990). Temperature, egg size, and development
of embryos and alevins of five species of Pacific salmon: a comparative
analysis.
\emph{Canadian Journal of Zoology},
\bold{68}, 1931--1940.

Sparks, M.M., Falke, J.A., Quinn, T.A., Adkinson, M.D.,
Schindler, D.E. (2019). Influences of spawning timing, water temperature,
and climatic warming on early life history phenology in western
Alaska sockeye salmon.
\emph{Canadian Journal of Fisheries and Aquatic Sciences},
\bold{76(1)}, 123--135.

Austin, C.C., Essington, T.E., Quinn, T.A. (2019). Spawning and emergence
phenology of bull trout \emph{Salvelinus confluentus} under differing
thermal regimes.
\emph{Canadian Journal of Fisheries and Aquatic Sciences},
\bold{94(1)}, 191--195.
}
\keyword{datasets}
