% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_check_temp.R
\name{plot_check_temp}
\alias{plot_check_temp}
\title{Plot imported data}
\usage{
plot_check_temp(data, dates, temperature, temp_min, temp_max)
}
\arguments{
\item{data}{imported data}

\item{dates}{date column}

\item{temperature}{temperature column}

\item{temp_min}{min temp on plot}

\item{temp_max}{max temp on plot}
}
\description{
Plot imported data
}
