% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenology.R
\name{plot_phenology}
\alias{plot_phenology}
\title{Plot phenology}
\usage{
plot_phenology(plot, style = "all", labels = TRUE)
}
\arguments{
\item{plot}{object output from predict_phenology}

\item{style}{different plot types: all, ef_cumsum, ef_daily}

\item{labels}{true or false, used to turn on and off the information at the top of plots}
}
\value{
ggplot2 object
}
\description{
Plot phenology
}
\examples{
library(hatchR)
plot_phenology(WI_hatch, style = "ef_cumsum") # shows a plot with just the ef cumulative sum values
plot_phenology(WI_hatch, style = "ef_daily") # shows a plot with just the ef daily values
plot_phenology(WI_hatch, labels = FALSE) # turns off the labeling for a cleaner figure
}
