% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenology.R
\name{plot_phenology}
\alias{plot_phenology}
\title{Plot phenology}
\usage{
plot_phenology(plot, style, labels)
}
\arguments{
\item{plot}{object output from predict_phenology}

\item{style}{different plot types (all, ef_cumsum, ef_daily)}

\item{labels}{true or false, used to turn on and off the information at the top of plots}
}
\value{
ggplot2 object
}
\description{
Plot phenology
}
\examples{
to come...
}
