% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenology.R
\name{plot_phenology}
\alias{plot_phenology}
\title{Plot phenology}
\usage{
plot_phenology(plot, style, labels)
}
\arguments{
\item{plot}{blah}

\item{style}{blah}

\item{labels}{blah}
}
\value{
blah
}
\description{
Plot phenology
}
\examples{
to come...
}
