% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenology.R
\name{plot_phenology}
\alias{plot_phenology}
\title{Visualize the phenology of fish development}
\usage{
plot_phenology(plot, style = "all", labels = TRUE)
}
\arguments{
\item{plot}{A list containing the output from \code{predict_phenology()}}

\item{style}{The style of the plot. A vector with possible values "all",
"ef_cumsum", "ef_daily". The default is "all".}

\item{labels}{Logical. If TRUE (default), labels are added to the plot.}
}
\value{
ggplot2 object
}
\description{
The function takes the output from \code{predict_phenology()}
and creates a ggplot2 object.
}
\examples{
library(hatchR)
#plot_phenology(WI_hatch, style = "ef_cumsum") # shows a plot with just the ef cumulative sum values
#plot_phenology(WI_hatch, style = "ef_daily") # shows a plot with just the ef daily values
#plot_phenology(WI_hatch, labels = FALSE) # turns off the labeling for a cleaner figure
}
