% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_phenology.R
\name{predict_phenology}
\alias{predict_phenology}
\title{Predict phenologt of fish based on temperature}
\usage{
predict_phenology(data, dates, temperature, spawn.date, model)
}
\arguments{
\item{data}{Dataframe with dates and temperature.}

\item{dates}{Date of temperature measurements.}

\item{temperature}{Temperature measurements.}

\item{spawn.date}{Date of spawning.}

\item{model}{Model to predict phenology.}
}
\value{
A list with the following elements:
}
\description{
A short description...
}
\examples{
# to come...
}
