% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_phenology.R
\name{predict_phenology}
\alias{predict_phenology}
\title{Predict phenology of fish}
\usage{
predict_phenology(data, dates, temperature, spawn.date, model)
}
\arguments{
\item{data}{Dataframe with dates and temperature.}

\item{dates}{Date of temperature measurements.}

\item{temperature}{Temperature measurements.}

\item{spawn.date}{Date of spawning.}

\item{model}{Model to predict phenology.}
}
\value{
A list with the following elements:
\itemize{
\item \code{ef.vals}: A numeric vector of each day's effective value.
\item \code{days2done}: A numeric vector of length 1; number of predicted days to hatch or emerge.
\item \code{ef.tibble}: An n x 4 tibble (n = number of days to hatch or emerge) with
the dates, temperature, effective values, and cumulative sum of the effective values.
\item \code{dev.period}: a 1x2 dataframe with the dates corresponding to when your
fish's parent spawned (input with \code{predict_phenology(spawn.date = ...)})
and the date when the fish is predicted to hatch or emerge.
}
}
\description{
A short description...
}
\examples{
# to come...

}
\references{
Sparks, M.M., Falke, J.A., Quinn, T.P., Adkinson, M.D., Schindler, D.E.
(2017). Influences of spawning timing, water temperature, and
climatic warming on early life history phenology in western
Alaska sockeye salmon.
\emph{Canadian Journal of Fisheries and Aquatic Sciences},
\bold{76(1)}, 123--135
}
