% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_temp.R
\name{summarize_temp}
\alias{summarize_temp}
\title{Summarize Temperature Data}
\usage{
summarize_temp(data, dates, temperature)
}
\arguments{
\item{data}{data.frame with dates and temperature}

\item{dates}{date column}

\item{temperature}{temperature column}
}
\value{
A data frame with summarized daily temperature values.
}
\description{
The \code{summarize_temp} function is used to summarize sub-daily temperature
measurements to obtain mean daily temperature.
}
\examples{
library(hatchR)
summarize_temp(data = crooked_river,
  dates = date,
  temperature = temp_c
  )
}
