% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woody_island.R
\docType{data}
\name{woody_island}
\alias{woody_island}
\title{Temperature Data from Lake Iliamna, Alaska}
\format{
Data frame
\describe{
\item{author}{Short article citation}
\item{species}{Species name}
\item{model}{Model number from B&M}
\item{dev.type}{Hatch or emergence}
\item{func}{character string desciption of the function}
}
}
\source{
To come...
}
\usage{
woody_island
}
\description{
Temperature Data from Lake Iliamna on Woody Island, Alaska
}
\examples{
  woody_island
}
\keyword{datasets}
